+++ src/frontend/apps/impress/src/core/auth/api/types.ts
@@ -8,4 +8,6 @@
 export interface User {
   id: string;
   email: string;
+  full_name: string;
+  short_name: string;
 }